add daily logger, fix secure for checkin and reset password

+++ fe-app/src/views/pages/admin/departments/Departments.tsx
@@ -43,8 +43,10 @@ const Departments: React.FunctionComponent = (): React.ReactElement => {
 
   const deletedepartment = async (id: number) => {
     try {
-      await departmentService.deleteDepartment(id);
-      getListDepartment();
+      if (window.confirm('Bạn có chắc chắn muốn thực hiện xóa ?')) {
+        await departmentService.deleteDepartment(id);
+        getListDepartment();
+      }
     } catch (error) {
       window.alert('Xảy ra lỗi khi xóa');
     }
